remove bulk prediction functionality
we have two options for accepting multiple predictions:

1) instructing the client to send a request for each prediction
2) allowing the client to pass a list of specs to predict for

the second option presents a problem when we get into the weeds
of GET request length limits. since specs can vary widely in
length when running json.dumps() on them, it would be difficult
to either instruct the client to limit their length, or to
somehow validate it.

when we were allowing bulk prediction in one GET request,
it was handled through an asyncio.gather() call. this approach
is approximately 2x faster than individual HTTP requests when
testing 5000 consecutive calls to the API (4s vs 8s). In this case,
8 seconds is not very long so it's not a problem, but we can revisit this
in the future if we run into performance issues.