Prediction API (#5)
* first prediction functionality

* 1

* add basic routing

* check whether predictions are lower than current allocation

* adjust conversions

* fix some bugs

* no black current_mapping

* readability

* allocate -> allocation

Co-authored-by: Alec Scott <scott112@llnl.gov>

* use GET endpoint rather than POST for allocation

* simplify validation logic

* bulk predictions run in parallel

* fmt changes

* cpu units now in cores rather than millicores

* request limit 8MB

* add variant filtering

* small fixes

* updated standard allocations

* Re-write the variant filtering logic.

Rather than grabbing the 4-5 rows that match based on other params (name, version for both pkg and compiler) and then filtering by variant, these queries are now split up into two. The program will first try to do an exact match on variants. If that doesn't work, it'll try to match based on expensive variants. This increases the amount of queries being done, but is a more robust system and will result in more matches.

* formatting

* formatting 😵‍💫

* logging -> logger usage

* remove unnecessary comment [ci skip]

* move validate payload for prediction into util for easier testing

* make a couple changes to prediction to make testing easier

* remove bulk prediction functionality

we have two options for accepting multiple predictions:

1) instructing the client to send a request for each prediction
2) allowing the client to pass a list of specs to predict for

the second option presents a problem when we get into the weeds
of GET request length limits. since specs can vary widely in
length when running json.dumps() on them, it would be difficult
to either instruct the client to limit their length, or to
somehow validate it.

when we were allowing bulk prediction in one GET request,
it was handled through an asyncio.gather() call. this approach
is approximately 2x faster than individual HTTP requests when
testing 5000 consecutive calls to the API (4s vs 8s). In this case,
8 seconds is not very long so it's not a problem, but we can revisit this
in the future if we run into performance issues.

* fix: remove unnecessary condition in validate_payload

* fix: remove custom GET length, 8K is sufficient for one url-encoded json spec

* fix: remove hash from GET payload

* fix: update docstring for allocation view

* allocation API now accepts spec string rather than JSON payload

* formatting fixes

* formatting fixes

* black is not being consistent :(

* fix: race condition where spec passed by reference was modified by `get_sample`

* `predict_single` -> `predict`

Co-authored-by: Alec Scott <scott112@llnl.gov>

* decouple prediction strategy from env variable

---------

Co-authored-by: Alec Scott <scott112@llnl.gov>