Add index for common prediction query
Previously, requests to the `/v1/allocate` endpoint took around 57ms. After adding a composite index that covers the complete spec and the most common query during the prediction algo, we get a large speedup to around 1.3ms per request. While more indices could be added, the returns were diminished and more risky as we don't want to negatively impact the INSERT speed for the jobs table.