Gracefully handle HTTP errors
closes #111

We were setting `raise_for_status` and not handling exceptions raised by aiohttp, which clogged up the logs with stack traces and env variables.

I did not add an automatic retry to HTTP requests because the issues in the cluster are not necessarily going to solve themselves in the scope of 5 retried requests.

I'll be dealing with the gitlab 500 errors in a different PR.